+++ test/bigtest/network/config.js
@@ -1,4 +1,45 @@
+import CQLParser from './cql';
+
 // typical mirage config export
-// http://www.ember-cli-mirage.com/docs/v0.4.x/configuration/
 export default function config() {
+  this.get('/configurations/entries', {
+    configs: []
+  });
+
+  this.get('/circulation/loans', {
+    loans: [],
+    totalRecords: 0,
+  });
+
+  this.get('/proxiesfor', {
+    proxiesFor: [],
+    totalRecords: 0,
+  });
+
+  this.get('/manualblocks', {
+    manualblocks: [],
+    totalRecords: 0,
+  });
+
+  // requests
+  this.get('/circulation/requests', ({ requests }, request) => {
+    if (request.queryParams.query) {
+      return requests.all();
+    } else {
+      return [];
+    }
+  });
+
+  // users
+  this.get('/users', ({ users }, request) => {
+    if (request.queryParams.query) {
+      const cqlParser = new CQLParser();
+      cqlParser.parse(request.queryParams.query);
+      return users.where({
+        barcode: cqlParser.tree.term
+      });
+    } else {
+      return [];
+    }
+  });
 }