+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 ## 1.5.0 IN PROGRESS
 
 * Adjust checkout columns width. Fixes UICHKOUT-457.
+* Add items awaiting pickup modal. Part of UICHKOUT-475.
 * Fix enforce manual patron blocks. Fixes UIU-792.
-
 
 ## [1.4.0](https://github.com/folio-org/ui-checkout/tree/v1.4.0) (2018-12-13)
 [Full Changelog](https://github.com/folio-org/ui-checkout/compare/v1.3.1...v1.4.0)